# api-docs.yaml
## @RequestMapping without explicit method
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: "http://localhost:8080"
  description: Generated server url
paths:
  /rest/p3:
    get:
      tags:
      - my-rest-controller
      operationId: p3
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    put:
      tags:
      - my-rest-controller
      operationId: p3_1
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    post:
      tags:
      - my-rest-controller
      operationId: p3_4
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - my-rest-controller
      operationId: p3_3
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    options:
      tags:
      - my-rest-controller
      operationId: p3_6
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    head:
      tags:
      - my-rest-controller
      operationId: p3_5
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    patch:
      tags:
      - my-rest-controller
      operationId: p3_2
      parameters:
      - name: param1
        in: query
        required: true
        schema:
          type: string
      - name: myheader1
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components: {}